[XNNPACK][QS8] torch.slice
Pull Request resolved: https://github.com/pytorch/pytorch/pull/105252


ghstack-source-id: 195068479

Supporting quantized slice operation. Currently there is no way to support the operator from short-term solution, so using long term QNNPACKQuantizer to quantize the model for slice.

Slice Operation is quantized by propagation. This means that the op can not be quantized by itself but will be quantized if its input is quantized.

Differential Revision: [D47487423](https://our.internmc.facebook.com/intern/diff/D47487423/)